Extract `SocketUtil` from `TcpWrangler`.
